Update release instructions and clean up update deps helper
