+++ tool/pubspec.yaml
@@ -1,8 +1,12 @@
-name: devtools_repo
+name: devtools_tool
 description: A repo management tool for DevTools.
+publish_to: none
 
 environment:
   sdk: '>=2.12.0 <4.0.0'
+
+executables:
+  repo_tool: repo_tool
 
 dependencies:
   args: ^2.1.0
